Apply better multidimensional array conventions
This is an attempte to resolve compiler errors on suwa's end, by using
more consistent 2D array access.

However, I have yet to test this on a Windows machine, and I cannot
compile on Linux until we merge PR#4.
Do not be surprised when this fails to compile! I am simply trying to
"nudge" it in the right direction towards a compilable state.